Move config-related stuff to a separate module
